new ui system (there still might be some bugs), little improvements and fixes

+++ src/engine/filer.h
@@ -17,7 +17,7 @@ EDirFilter operator|(EDirFilter a, EDirFilter b);
 
 class Filer {
 public:
-	static int CheckDirectories();
+	static byte CheckDirectories();
 	static bool ReadTextFile(string file, vector<string>& lines);
 	static bool WriteTextFile(string file, const vector<string>& lines);
 	static vector<fs::path> ListDir(fs::path dir, EDirFilter filter, const vector<string>& extFilter=vector<string>());
